+++ src/main/java/frc/robot/RobotContainer.java
@@ -7,12 +7,12 @@ package frc.robot;
 import frc.robot.Constants.OperatorConstants;
 import frc.robot.Constants.DashboardConstants;
 import frc.robot.Constants.DrivetrainConstants;
-import frc.robot.subsystems.EndEffector.EndEffector;
 import frc.robot.subsystems.drivetrain.Drivetrain;
+import frc.robot.subsystems.endEffector.EndEffector;
+import frc.robot.subsystems.intakeRamp.Ramp;
 import frc.robot.util.Elastic;
 import frc.robot.Constants.OperatorConstants.GamepieceMode;
 import frc.robot.commands.StubbedCommands;
-import frc.robot.subsystems.IntakeRamp.Ramp;
 import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.RobotBase;
 import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
